+++ NBALigaSimulation/Shared/Dtos/Player/PlayerGameStatsDto.cs
@@ -32,8 +32,8 @@
         public int Pf { get; set; }
         public int Pts { get; set; }
         public int Trb { get; set; }
-        public double CourtTime { get; set; }
-        public double BenchTime { get; set; }
+        public decimal CourtTime { get; set; }
+        public decimal BenchTime { get; set; }
         public int Energy { get; set; }
     }
 }
